requirements.txt
click==8.1.4
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pyparsing==3.1.0
six==1.16.0
Werkzeug==2.3.5
python-jobspy==1.1.63
APScheduler==3.10.4
pinecone-client==5.0.1
openai==1.39.0
langchain-community==0.2.11
llama-index-core==0.10.61
llama-parse==0.4.9
llama-index-readers-file==0.1.32
python-dotenv